requirements.txt
python-dotenv==1.0.0
instaloader==4.10
selenium~=4.12.0